Adding flag to force refresh during mirror
Looking at the usage in lims/infrastructure/base-images, I believe I set up my updates on line 35 correctly, but my gitlab/shell-fu is not super strong.
Intent is that if the new flag "FORCE_REFRESH" is not present, no differences will occur and the pull of the image will not happen if the image is already seen in the registry. Otherwise, it would update the image in the registry for the requested tag.